migrations files update

+++ migrations/8_create_retailers_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS retailers(
     id SERIAL PRIMARY KEY,
-    category_id INTEGER NOT NULL,
+    category_id SMALLINT NOT NULL,
     FOREIGN KEY(category_id) REFERENCES retailer_categories(id),
     retailer_name VARCHAR(25) NOT NULL
 );
